+++ .test/test.sh
@@ -103,3 +103,11 @@ docker run \
     --shm-size=2G \
     taconsol/sakuli:${1:-latest}
 
+# OCR run
+docker run \
+    --rm \
+    -e SAKULI_LICENSE_KEY=${SAKULI_LICENSE_KEY} \
+    -e SAKULI_TEST_SUITE=/sakuli-project/e2e-ocr \
+    -v $(pwd)/e2e:/sakuli-project \
+    --shm-size=2G \
+    taconsol/sakuli:${1:-latest}