+++ run.sh
@@ -1,9 +1,7 @@
-#!/usr/bin/env bash
-
 docker run \
     --rm \
     -it \
     -e SAKULI_LICENSE_KEY=${SAKULI_LICENSE_KEY} \
-    -p 5902:5901 \
-    -p 6902:6901 \
+    -p 5901:5901 \
+    -p 6901:6901 \
     taconsol/sakuli